+++ src/components/Pages/AdminDashboard/AdminDashboard.jsx
@@ -8,6 +8,7 @@ import Status from "./Status";
 import SalesByStaff from "./SalesByStaff";
 import ModeOfPayment from "./ModeOfPayment";
 import TotalSales from "./TotalSales";
+import TargetProgressCard from "./Target";
 
 const AdminDashboard = () => {
     const [selectedDate1, setSelectedDate1] = useState(new Date());
@@ -126,120 +127,9 @@ const AdminDashboard = () => {
                     </div>
 
                     {/* Row 3: Sales by Staff, Top 5 Customers, Top 5 Services */}
-                    <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-6 mt-6">
+                    <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-2 gap-6 mt-6">
+                        <TargetProgressCard />
                         <SalesByStaff />
-                        {/* Top 5 Customers */}
-                        <div className="bg-white shadow-md p-6 rounded-lg">
-                            <div className="flex justify-between items-center mb-4">
-                                <h3 className="text-lg font-semibold text-gray-700 mb-4">
-                                    Top 5 Customers
-                                </h3>
-                                <select
-                                    value={period}
-                                    onChange={(e) => setPeriod(e.target.value)}
-                                    className="border border-gray-300 text-gray-700 rounded-lg p-1 text-sm"
-                                >
-                                    <option value="Month">Month-to-date</option>
-                                </select>
-                            </div>
-                            {topCustomers.length > 0 ? (
-                                <ul>
-                                    {topCustomers.map((customer, index) => (
-                                        <li
-                                            key={index}
-                                            className="flex justify-between py-2 text-gray-700 border-b last:border-b-0"
-                                        >
-                                            <span>
-                                                {index + 1}. {customer.name}
-                                            </span>
-                                            <span>
-                                                Rs.
-                                                {customer.revenue.toLocaleString()}
-                                            </span>
-                                        </li>
-                                    ))}
-                                </ul>
-                            ) : (
-                                <div>
-                                    {/* Red warning message */}
-                                    <p className="text-red-600 font-light mb-2">
-                                        *No data available for the selected
-                                        period.
-                                    </p>
-                                    {/* Placeholder rows */}
-                                    <ul>
-                                        {[1, 2, 3, 4, 5].map((_, index) => (
-                                            <li
-                                                key={index}
-                                                className="flex justify-between py-2 text-gray-400 border-b last:border-b-0 italic"
-                                            >
-                                                <span>
-                                                    {index + 1}. Customer
-                                                </span>
-                                                <span>Rs.__</span>
-                                            </li>
-                                        ))}
-                                    </ul>
-                                </div>
-                            )}
-                        </div>
-
-                        {/* Top 5 Services */}
-                        <div className="bg-white shadow-md p-6 rounded-lg">
-                            <div className="flex justify-between items-center mb-4">
-                                <h3 className="text-lg font-semibold text-gray-700 mb-4">
-                                    Top 5 Services
-                                </h3>
-                                <select
-                                    value={period}
-                                    onChange={(e) => setPeriod(e.target.value)}
-                                    className="border border-gray-300 text-gray-700 rounded-lg p-1 text-sm"
-                                >
-                                    <option value="Month">Month-to-date</option>
-                                </select>
-                            </div>
-                            {topServices.length > 0 ? (
-                                <ul>
-                                    {topServices.map((service, index) => (
-                                        <li
-                                            key={index}
-                                            className="flex justify-between py-2 text-gray-700 border-b last:border-b-0"
-                                        >
-                                            <span>
-                                                {index + 1}. {service.name}
-                                            </span>
-                                            {/* <span>Occured: {service.occurrences.toLocaleString()}</span> */}
-                                            <span>
-                                                Rs.
-                                                {service.revenue.toLocaleString()}
-                                            </span>
-                                        </li>
-                                    ))}
-                                </ul>
-                            ) : (
-                                <div>
-                                    {/* Red warning message */}
-                                    <p className="text-red-600 font-light mb-2">
-                                        *No data available for the selected
-                                        period.
-                                    </p>
-                                    {/* Placeholder rows */}
-                                    <ul>
-                                        {[1, 2, 3, 4, 5].map((_, index) => (
-                                            <li
-                                                key={index}
-                                                className="flex justify-between py-2 text-gray-400 border-b last:border-b-0 italic"
-                                            >
-                                                <span>
-                                                    {index + 1}. Service
-                                                </span>
-                                                <span>Rs.__</span>
-                                            </li>
-                                        ))}
-                                    </ul>
-                                </div>
-                            )}
-                        </div>
                     </div>
                 </div>
             </div>
